Add functions to dump contents of index maps.
Adds dump_as_array() and dump_as_list() methods to map classes that take a fd
and dump the contents of the index maps to disk. I am not sure whether thats
the right interface, but here it is. We can always change it later.
There is a new osmium_index program in the example directory, that allows
looking up elements in these indexes and dumping them in readable format.
osmium_toogr now dumps its location index to disk.