Support for a sort of OSM database on disk.
This includes a handler (DiskStore) that writes the serialized OSM data to disk
and creates indexes by id to that data. The ObjectRelations handler keeps
mappings from child object ID to parent objects IDs. And there are new index
types to support these. And osmium_serdump to tie everything together and test
it.