Use std::pair<TKey, TValue> as element_type in indexes.
Instead of osmium::index::detail::element_type. This makes it easier to use the
STL map/multimap classes which use that type internally. It doesn't really
change much in it's use. Instead of 'key' and 'value' members, we have 'first'
and 'second'. And we have to use a lambda comparison function when we search
inside a container instead of using operator< on element_type.
This commit also implements a few dummy functions needed to make the interface
of the multimap classes the same.