Add (optional) memory management to Buffer class.
The Buffer class can now be used in two ways: With internal memory management
or with external memory management. In the PBF and XML readers the internal
memory management is used now which obviates the need for manual memory
management in typical Osmium programs. This makes the use of the Buffer class
easier while still allowing the full flexibility when using external memory
management.