Add support for only reading some types of data from file.
It is quite common that some program reads an OSM file but is only interested
in, say nodes, but not ways and relations. Of course it costs time to parse
things that are not needed, so there are now flags given to the Reader open()
call which objects you are interested in. Note that it will still read the
whole file and, at least, superficially parse it, because it can't know whether
one little node will be at the end of the file. If you are only interested in
the header and nothing else, you can set this flag to "nothing" and it will not
read the entire file.