Make Node, Way, Relation constructors private.
Those objects should never be created by themselves but only by the
corresponding Builder classes. The constructors have now been declared private
and the Builder class declared as a friend to enforce this. Some tests needed
to be adapted, too. And while we are at it those tests are made a bit easier
by introducing some helper functions.