Cleanup of PBF input code.
The BlobParser class is initalized in one thread where it reads data into a
buffer and then does the actual parsing in another thread. Due to the way
std::function works, which is at the base of the thread work queue, the
BlobParser class (or, to be more exact, a derived class) is copied and not
moved when it is given to the new thread. There was some problematic code
in there using a plain pointer and a unique_ptr which has now been cleaned
up to use a shared_ptr to make the copy reasonably cheap and still have
exception-correct-code.