Change thread pool implementation.
First, there is only one thread pool now implemented as singleton. Every part
of Osmium that needs worker threads uses this single pool.
Second, we are using a function_wrapper helper class to wrap the tasks in
the worker queue. Unlike the std::function used before, the function_wrapper
is movable. We use packaged_tasks to allow return values given back with
futures (which are not used yet in this commit).