XML output redone.
This changes the XML writer to run in several threads using the worker
thread pool.
This also removes the dependency on libxml2 for XML output. The XML is written
"manually". There are only a few places where we need to escape data and
we only deal with UTF-8 data and we only write UTF-8 encoded files anyway,
so this is easy to do.