Add first stab at multipolygon code.
This commit adds all the scaffolding needed before we can do the actual
multipolygon assembling. So it does the collection of all the members of
multipolygon relations, deciding when we have all the data needed and then
calling the actual assembly. The assembly code currently is only a placeholder
that does as very simplistic (and often wrong) multipolygon assembly. There are
two test programns that dump out the created areas (osmium_mpdump) and create
OGR geometries from the assembled areas (osmium_toogr2).