Improved ProblemReporter
Because all problems are related to an object we are currently assembling into
an area, the problem reporter will get the type and ID of the object once and
can then re-use it. This reduces code duplication.
Also in this commit the ProblemReporterException uses functionality of the
ProblemReporterStream class and the ProblemReporterStream class was simplified,
all to allow more code reuse.