Remove init() and before_*/after_* functions in handlers.
This was a nice idea, but they create more problems than they solve. What
happens exactly if objects are not ordered according to type? What happens when
we read several files into the same handler? Only a few handlers actually needed
those calls and they could be rewritten to work without them.

The done() function was retained for the moment because the multipolygon
collection code needs it. After that code has been refactored we can get rid of
the done() function, too.