Retire done() handler call.
Rename it to flush(). It should only be used to flush any data that the handler
might have cached. Final cleanup should be done in handler destructor or other
handler specific functions. This way apply() can be called several times on the
same handlers, for instance for each of several buffers, and after each apply
flush() will be called but not the cleanup.

The multipolygon code has also been changed to take this into account.