Allow dynamic setting of node location cache type.
The constructor of LocationHandler can now be called with the type of the node
cache we want to use as a string. Currently supported are "sparsetable"
(default), "stlmap", "array" (Linux only), and "disk". The "disk" cache type
can take an optional file name argument for opening an existing file:
"disk,filename".