First stab at a "flexible" reader class that handles mp assembly automatically.
The FlexReader will automatically read the input file several times if needed
to assembly multipolygons. Also the read() function will return a vector of
Buffer objects instead of a single Buffer. When assembling multipolygons this
vector will usually contain two buffers, one with the way and/or relation
objects, one with the completed area objects.