Add way to dynamically choose node location index.
Until now you had to decide at compile time which node location index type you
wanted to use. It is now possible to decide this at run-time (for the price
of a small overhead). There is code to test this and the osmium_toogr example
has also been changed to take advantage of this.