Fix warnings on Windows.
On Windows unsigned long is only 32bit and not the same as size_t. ZLib takes
unsigned long for size arguments. This commit adds a cast_with_assert to fix
the warning and documents in the functions description that there is a size
limit on these functions. In libosmium this isn't a problem, because those
functions are only called from the PBF parser/writer and we know the blobs are
small enough by definition of the OSM PBF format.