Changed add_user() and add_role() in builders. Add add_member().
All these functions now have a version that takes a const char* and a length.
The length is the length without the 0-termination byte. This is consistent
with strlen() and it allows to work with strings that are not 0-terminated.
Note that this is an incompatible change to the API, but because this code
is usually only used in readers, it should be okay (fingers crossed).