New PBF reader and writer based on protozero.
This is a complete rewrite of the code for reading and writing OSM PBF files.
It doesn't use the Google protobuf library and it doesn't use the
OSMPBF/OSM-Binary library any more. Instead is uses the protozero lightweight
protobuf header library which is included in the code. Not only does the new
code have less dependencies, it is faster and more robust.