use less variables when defining the test environment
Define a CMAKE_PREFIX_PATH and let CMake look into "include" and "lib"
subdirectories itself. Only ZLIB_LIBRARY is still explicitely passed as it uses
a name which is not included in FindZLIB.cmake, as well as BZIP2_LIBRARIES
because of a strange problem with CMake not being able to find it if the path
contains backslashes.