Move internal buffer from OutputIterator into Writer.
The iterator becomes much smaller (which is important because iterators get
copied around a lot). The need for the smart_ptr<Buffer> is gone which should
speed things up. And it saves us the akward flush() call on the OutputIterator
that was easy to forget. The price we pay for that is an extra buffer in the
Writer (which is already a larger class that can't be copied, so not a big
deal). And if the buffer is not used it will not be initialized, so it doesn't
need any heap allocation. Some functions have been marked deprecated.