When the buffer isn't big enough, even when empty, to reserve the space that's being requested then it shouldn't reserve it anyway.
This fixes one bug, in which a fixed-size (not auto-growing)
buffer would call the `m_full` function and then return without
checking that `size < m_capacity`. It also makes the buffer more
flexible, as previously it wasn't possible to grow the buffer if a
"full" function was provided.

With this patch, it now attempts to flush the buffer and, if there
still isn't enough space, it can grow (depending on whether that
option is set). This allows the common case of flushing with a
pre-allocated small fixed-size buffer, while also allowing
larger-than-expected objects to be inserted into the buffer.