Bugfix: Improved segment intersection function.
Uses more integer arithmetic. This way the comparisons in the if-clause will
actually work and it should be faster. Documented that the whole function
will not work for very long segments (longer than about half the planet),
which shouldn't happen in real data anyway.

See 135. This fixes the testdata-multipolygon test failures.