Add support for reading/writing XML/PBF files with locations on ways.
Usually only node have locations and the ways refer to those locations via the
IDs of the nodes. This introduces the "locations_on_ways" option for XML and
PBF files. If it is set, locations will be written on the ways, too, if they
are available in the way objects. When reading XML or PBF files and there are
locations available on the ways they will be contained in the way objects. This
is a non-standard extension.