Fix for possible threading problem.
The Queue implementation had a mechanism to know whether it was in use or in
the process of being shut down using the m_done atomic bool. But this
interfered with the externally used mechanism of putting special "end of data"
markers in the queue which could lead to a situation where queue readers
would not shutdown properly.