Completely new algorithm for assembling multipolygons.
Assembling of multipolygons was too slow and did have several errors. The new
code is superfast. It needs only a few minutes to assemble all multipolygons
on the planet.

Some problems in the OSM data can be fixed, but if segments intersect, this
code will not build the multipolygon. This can be improved upon later.

The always tries to generate valid geometries with the outer rings oriented
counter-clockwise and inner rings clockwise according to the Simple Feature
Specification.

The code will now also generate lots of statistics and better error reports
to show where the problems in the multipolygons are.

There are some bugs left in this version. Some very complex multipolygons with
multiple locations where several rings touch in one point will not be built
correctly, but this code is still a huge improvement over existing code.