Automatically set correct SRID when creating geometries.
This works for EWKB and GEOS geometry types. When the IdentityProjection is
used (default), the SRID is set to 4326. When MercatorProjection is used,
SRID 3857 is used. For other projections their respective EPSG code is used.

Note that this changes the behaviour of libosmium when creating GEOS
geometries. Before we created them with -1 as SRID unless set otherwise.
Manual setting of the SRID on the GEOSGeometryFactory is now deprecated.

OGR, JSON, and WKT geometries do not have SRIDs attached. EWKT is currently
not supported.

See #138.