Bugfix: PBF String table corruption when there are many strings.
When there are many and/or long strings in tag keys and values and/or user
names and/or relation roles, the string table inside a PBF block would
overflow. I have never seen this happen for normal OSM data, but that doesn't
mean it can't happen. The result is that the strings will all be mixed up, keys
for values, values for user names or whatever.