In debug mode check that Builders are correctly called and destructed.
Builders must be called and destructed in the right order. This adds
some code to the Buffer and some asserts in the Builder that (when
NDEBUG is not set) will check everything is done correctly. If you
get assertion failures after this update, check that Buffer::commit()
is only called after all Builder objects are destructed. In addition
only one sub-builders can be open at any time.