Check GEOS version is <= 3.5.
The current code only works with GEOS up to 3.5, because there were
changes in the C++ API of GEOS that we are using. Using the C++ API was
probably a bad idea anyway, because the GEOS project doesn't consider
the C++ API to be a stable external API. It would be better to use the
stable C API of GEOS, but until somebody needs that and is willing
to implement it, we will not have that.