Use condition_variable to tell producer when thread queue is not full.
Until now we slept in the producer if the queue was full, waking up
occasionally to check whether we could go on. This turned out to be too
simple. The wait time depends on the speed of the consumers and tuning
the wait time didn't give satisfactory results in all cases. The new
implementation is more complicated, but benchmarks show that it gives
better performance in a variety of uses.