Bugfix: Terminate called on full buffer.
A full non-auto-growing buffer could lead to the program terminating
instead of a thrown exception. Reason was that the number of bytes
written to the buffer was not tracked properly when appending tag keys
and values to the buffer. The destructor on the TagListBuilder would
then throw another exception which terminates the program. This should
fix this problem by keeping the number of bytes written to the buffer
correctly.

See #189.