Add compiler/libc checks for disabling std::regex use in StringMatcher.
std::regex isn't implemented properly in glibc++ (before the version
delivered with GCC 4.9) and libc++ before the version 3.6, so the use is
disabled by these checks.
Checks for GLIBC were based on
https://stackoverflow.com/questions/12530406/is-gcc-4-8-or-earlier-buggy-about-regular-expressions
Checks for libc++ are simply based on compiler defines. This is probably
not optimal but seems to work for now.