Rewrite switch statements to avoid compiler warnings.
We have a problem with switch statements: Some compiler versions warn
when a switch doesn't handle all options, so we need them to be
complete. But if they are complete, there is never a case where the
code falls out of the switch, so the return at the end of the function
is never called (which some compilers don't like), but if there is no
return at the end of the function other compiler versions complain. The
only way to handle this is to pick one of the cases, make it into a
default case that drops out of the switch and then do the return
outside the switch.