Check order of input data in relations collector.
The relations collector can not deal with history data or a changes
file. This was documented as a requirement, but often lead to problems,
because this was ignored by users. So it now checks that the input data
it gets is ordered and throws an osmium::out_of_order_error exception
otherwise.

See osmcode/pyosmium#32.