Change eof behaviour of Reader.
The Reader::read() function will now always return an invalid buffer
exactly once to signal eof. This even applies when
osm_entity_bits::nothing was set in which case the first call to that
function will return the invalid buffer and set the file to status eof.
In other words, if osm_entity_bits::nothing was set, the file just
looks empty.