Enable use of user-provided thread::Pool instance
Pool instance can be configured with user-specified number of threads
and passed to Reader and Writer as optional argument of constructors.
If no Pool is specified, default instance is used and configured with
number of threads based on actual number of cores on the system.

Advantages:
* Allows to use Pool configured with user-preferred number of threads.
* If libosmium used in a shared library, this helps to prevent potential deadlock
during Pool shutdown at (after) shared library unloading (see #213 for details).