New flex_mem auto-sizing index.
This index type will work with small and large files (including the
full planet). Depending on the number of Ids in the index it will
use a sparse or dense array. It usually starts out with a sparse
index and after it gets some data will decide to switch to the dense
index if the Id space is dense enough for it to make sense.