Changes for consistent ordering of OSM data.
OSM data can come in any order, but usual OSM files are ordered by
type, ID, and version. These changes extend this ordering to negative
IDs which are sometimes used for objects that have not been uploaded to
the OSM server yet. The negative IDs are ordered now before the
positive ones, both in order of their absolute value.

So the ordering is:
* node -1, node -2, ..., node 1, node 2, ...,
* way -1, way -2, ..., way 1, way 2, ...,
* relation -1, relation -2, ..., relation 1, relation 2, ...

This is the same ordering as JOSM uses.