Introduce special "blackhole" file format.
This can only be used for writing. It will throw away all the data
without even encoding it. This is useful in some cases, for instance
when you are benchmarking. Unlike writing to /dev/null which will
encode the data before throwing it away, the "blackhole" file type
doesn't have any overhead.