Add tests and remove assumed nesting object from RapidGeoJSON geom implementation
Previously the implementation of this assumed that the caller
had already opened a new JSON object in the provided stream and
that the requested geometry would be embedded inside of that.

This works well for the assumed case (Using this FactoryImpl to
generate just the 'geometry' key of a GeoJSON feature) but is less
flexible since the generated JSON can't stand on its own.

Hopefully this implementation allows for both use cases to be supported.

* Include a test showing how to reset and re-use the RapidJSON Stream