Make util namespace inline.
It was always a bit confusing that some of the util classes and
functions are directly in the osmium namespace and some are in
osmium::util. This change makes the util namespace inline, which allows
all util classes and functions to be addressed directly in the osmium
namespace while keeping backwards compatibility.