Change the way IDs are parsed from strings.
No asserts are used any more but checks are done and an exception is
thrown when IDs are out of range.

This also changes the way negative values are handled. The value -1 is
now always accepted for all IDs and returned as 0.

This deprecates the string_to_user_id() function, use string_to_uid()
instead which returns a different type.