Add Timestamp::to_iso_all() function which also outputs the 0 value.
In Osmium the 0 value of the Timestamp is used to denote the "invalid"
Timestamp, and its output using the `to_iso()` function is the empty
string. But this is the wrong output for OSM XML files, where a
timestamp that's not set should still be output as
1970-01-01T00:00:00Z. This commit introduces a new `to_is_all()`
function which will do this and uses that function in the XML writer.

See #254.