The apply function can now also take lambdas.
Writing handlers for small tasks is sometimes a bit annoying, lots of
boilerplate. This commit enables an alternative: Instead of a full
handler, the apply function can now be called with a lambda (or object
of a class with an operator() function). Depending on the type of
the parameter, this is called as the node, way, relation, area, or
changeset callback. See test/t/handler/test_apply.cpp for some
examples.

Closes #181.