Bugfix: Do not propagate C++ exception through C code.
We are using the Expat XML parser, a C library. It calls callbacks in
our code. When those callbacks throw, the exception was propagated
through the C code. This did work in the tests, but that behaviour
isn't guaranteed (C++ standard says it is implementation defined).
This commit fixes this by catching the exception and rethrowing
it later.

Fixes #287.