Also support API of modern PROJ library (version 6 and above)
Set the OSMIUM_USE_PROJ_H or OSMIUM_USE_PROJ_API_H to choose which API
version of the PROJ library should be used. For versions up to 5 use
OSMIUM_USE_PROJ_API_H, for version 6 both will work, for versions 7 and
above use OSMIUM_USE_PROJ_H.

If you are using cmake/FindOsmium.cmake, the right defines will be
chosen for you: If proj_api.h is found the old API will be used, if not,
it looks for proj.h and uses the new API.

This deprecates some functionality which can't be supported in the new
API, namely the osmium::geom::CRS class, the osmium::geom::transform()
function and the osmium::geom::Projection::proj_string() member
function.

See #295