Make timer class member functions non-static again
Otherwise this will get us into problems with VS which spits out a
warning C4101 that we are using an unreferenced local variable.

This case is a bit tricky because the code is intended to be compiled
away to nothing, but if OSMIUM_WITH_TIMER is set there should be
something there.

See https://github.com/osmcode/osmium-tool/issues/203