Workaround for GCC in C++17 mode
The line

static TSubitem subitem{};

is seen by GCC in C++17 mode (but not in C++11 or C++20 mode) as using a
move constructor which is not available for the classes used as
TSubitem. So this fails. But without the {} it works.

It shouldn't make any difference either way, because as a static
variable subitem will be zero-initialized, which is the right thing to
do here.