Make bzip2 unit tests pass on musl-based systems
Other than glibc, the musl standard C library does not check the
validity of file descriptors passed to `fdopen()`, so the constructor
of the (de)compressor does not throw an excption on musl-based systems.
Rather, an exception is thrown later, when the opened file is being used.
According to the musl team, the current musl behavior conforms to the POSIX
specification, and application code should be made agnostic to whether
or not the standard C library happens to validate file descriptors
passed to `fdopen()`.

For the discussion of `fdopen()` filedescriptor validation, see this
thread on the musl mailing list:

* https://www.openwall.com/lists/musl/2022/08/04/1
* https://www.openwall.com/lists/musl/2022/08/04/2
* https://www.openwall.com/lists/musl/2022/08/04/3

After this change, all libosmium tests (including those for bzip2) are
passing on Alpine Linux, which uses musl as its standard C library.

Fixes https://github.com/osmcode/libosmium/issues/353.