Fix problem when reading bz2 files
This fixes a case where libz2 tells us that there is no data any more
(BZ_STREAM_END), but we are not at the end of file, so there could be
more data. But then there isn't (nunused==0). It seems this doesn't
happen very often, but it does happen occasionally.

Fixes #373