Extend RelationsMapIndex to work with 64bit IDs
The RelationsMapIndex worked 32bit IDs only. That's totally fine for OSM
data because there aren't that many relations. But if you have some
other data in OSM formats this is not necessarily the case. The reason
to use 32bit IDs only is because you need only half the memory.

This commit changes the implementation that it uses 32bit IDs as long as
that's enough but switches to 64bit IDs if necessary. So for 32bit IDs
the impact is minimal while 64bit IDs are still possible.